Fixed a typo in AFIncrementalStore.m
Fixed what I'm 90% sure is a typo.
You've set `objectID` to nil, so `objectID.entity.name` will also be nil. I'm pretty sure this should be `entity.name`.
I noticed this while reading through the code trying to understand how it deals with object ids, so I can't attest to a problem caused by this or if this change will break anything. Hope it's helpful, tho!