Fix empty NSSaveChangesRequest notifications.
The original instance of NSSaveChangesRequest was passed to both
`AFIncrementalStoreContextWillSaveRemoteValues` and
`AFIncrementalStoreContextDidSaveRemoteValues`.

The probleme here was that CoreData removes the objects from the sets
referenced in the original request as they are saved. This means that
from a notification handler for
`AFIncrementalStoreContextDidSaveRemoteValues` you are provided with an
`NSSaveChangesRequest`instane whose sets are either empty or nil,
leaving you unable to get a reference on the managed objects.

This is fixed here by creating a new NSSaveChangesRequest with copies of
the original sets.